requirements.txt
python-dotenv
google-generativeai
google-cloud-bigquery
google-cloud-secret-manager
fastapi
uvicorn
gunicorn
